# Conflicts:
#	src/main/java/com/avada/myHouse24/config/SecurityConfig.java
#	src/main/java/com/avada/myHouse24/config/WebSocketConfig.java
#	src/main/java/com/avada/myHouse24/repo/FlatRepository.java
#	src/main/java/com/avada/myHouse24/services/impl/FlatServiceImpl.java
#	src/main/java/com/avada/myHouse24/util/ImageUtil.java
#	src/main/resources/application.yaml
#	src/main/resources/templates/admin/user/add.html
#	src/main/resources/templates/blocks/menu.html

+++ src/main/java/com/avada/myHouse24/services/impl/AdditionalServiceImpl.java
@@ -1,8 +1,15 @@
 package com.avada.myHouse24.services.impl;
 
+import com.avada.myHouse24.entity.House;
 import com.avada.myHouse24.repo.AdditionalServiceRepository;
+import com.avada.myHouse24.repo.CounterDataRepository;
+import com.avada.myHouse24.repo.InvoiceAdditionalServiceRepository;
+import com.avada.myHouse24.repo.TariffAdditionalServiceRepository;
 import com.avada.myHouse24.services.AdditionalService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,6 +18,9 @@ import java.util.List;
 @RequiredArgsConstructor
 public class AdditionalServiceImpl implements AdditionalService {
     private final AdditionalServiceRepository additionalServiceRepository;
+    private final InvoiceAdditionalServiceRepository invoiceAdditionalServiceRepository;
+    private final TariffAdditionalServiceRepository tariffAdditionalServiceRepository;
+    private final CounterDataRepository counterDataRepository;
     @Override
     public List<com.avada.myHouse24.entity.AdditionalService> getAll() {
         return additionalServiceRepository.findAll();
@@ -30,6 +40,7 @@ public class AdditionalServiceImpl implements AdditionalService {
 
     @Override
     public void deleteById(long id) {
+        if(isUses(id))return;
         additionalServiceRepository.deleteById(id);
     }
 
@@ -41,5 +52,21 @@ public class AdditionalServiceImpl implements AdditionalService {
     @Override
     public com.avada.myHouse24.entity.AdditionalService getByName(String name) {
         return additionalServiceRepository.findByName(name).get();
+    }
+    public List<com.avada.myHouse24.entity.AdditionalService> forSelect(int page, int pageSize, String search) {
+        Pageable pageable = PageRequest.of(page - 1, pageSize);
+        Page<com.avada.myHouse24.entity.AdditionalService> additionalPage;
+
+        if (search != null && !search.isEmpty()) {
+            additionalPage = additionalServiceRepository.findByNameContainingIgnoreCase(search, pageable);
+        } else {
+            additionalPage = additionalServiceRepository.findAll(pageable);
+        }
+
+        return additionalPage.getContent();
+    }
+    public Boolean isUses(long id){
+        if(counterDataRepository.existsByAdditionalService_Id(id) || tariffAdditionalServiceRepository.existsByAdditionalService_Id(id) || invoiceAdditionalServiceRepository.existsByAdditionalService_Id(id)) return true;
+        return false;
     }
 }
